# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/properties": {
      "post": {
        "operationId": "CreatePropertyController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyRequestDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new property",
        "tags": [
          "Properties"
        ]
      },
      "get": {
        "operationId": "GetPropertiesController_handle",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "VACANT",
                "LISTED",
                "OCCUPIED",
                "MAINTENANCE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of properties"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List properties for the authenticated manager",
        "tags": [
          "Properties"
        ]
      }
    },
    "/properties/{id}": {
      "get": {
        "operationId": "GetPropertyByIdController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Property UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property details"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get property details by ID",
        "tags": [
          "Properties"
        ]
      },
      "put": {
        "operationId": "UpdatePropertyController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Property UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update property details",
        "tags": [
          "Properties"
        ]
      },
      "delete": {
        "operationId": "DeletePropertyController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Property UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Property deleted"
          },
          "404": {
            "description": "Property not found"
          },
          "409": {
            "description": "Property has active lease"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a property",
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/expenses": {
      "post": {
        "operationId": "CreateExpenseController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "propertyId",
                  "category",
                  "amount",
                  "description",
                  "expenseDate"
                ],
                "properties": {
                  "propertyId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "category": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "expenseDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "maintenanceRequestId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense created"
          },
          "404": {
            "description": "Property not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new expense",
        "tags": [
          "Expenses"
        ]
      },
      "get": {
        "operationId": "GetExpensesController_handle",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "propertyId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Expenses list"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get expenses with filters",
        "tags": [
          "Expenses"
        ]
      }
    },
...
    "/expenses/{id}/receipt": {
      "post": {
        "operationId": "UploadExpenseReceiptController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Expense UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fileName",
                  "contentType"
                ],
                "properties": {
                  "fileName": {
                    "type": "string"
                  },
                  "contentType": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload URL generated"
          },
          "404": {
            "description": "Expense not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate upload URL for expense receipt",
        "tags": [
          "Expenses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mobile API",
    "description": "Auto-generated API docs with Swagger",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePropertyRequestDTO": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "123 Main St"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "zipCode": {
            "type": "string",
            "example": "10001"
          },
          "propertyType": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "CONDO",
              "TOWNHOUSE",
              "STUDIO"
            ],
            "example": "APARTMENT"
          },
          "bedrooms": {
            "type": "number",
            "example": 2
          },
          "bathrooms": {
            "type": "number",
            "example": 1.5
          },
          "sqft": {
            "type": "number",
            "example": 1200
          },
          "rentAmount": {
            "type": "number",
            "example": 2500
          },
          "description": {
            "type": "string",
            "example": "Spacious 2BR apartment with city views"
          }
        },
        "required": [
          "address",
          "city",
          "state",
          "zipCode",
          "propertyType",
          "bedrooms",
          "bathrooms",
          "rentAmount"
        ]
      },
      "UpdatePropertyRequestDTO": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "456 Oak Ave"
          },
          "city": {
            "type": "string",
            "example": "Los Angeles"
          },
          "state": {
            "type": "string",
            "example": "CA"
          },
          "zipCode": {
            "type": "string",
            "example": "90001"
          },
          "propertyType": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "CONDO",
              "TOWNHOUSE",
              "STUDIO"
            ],
            "example": "APARTMENT"
          },
          "bedrooms": {
            "type": "number",
            "example": 3
          },
          "bathrooms": {
            "type": "number",
            "example": 2
          },
          "sqft": {
            "type": "object",
            "example": 1500,
            "nullable": true
          },
          "rentAmount": {
            "type": "number",
            "example": 3000
          },
          "description": {
            "type": "object",
            "example": "Updated description",
            "nullable": true
          }
        }
      },
...
    }
  }
}